feature: report peer server down when it passively exits
